backend/requirements.txt
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
celery[redis]>=5.3.0,<6.0
redis>=5.0.0,<6.0
sqlalchemy[asyncio]>=2.0.25,<3.0
asyncpg>=0.29.0,<1.0
psycopg2-binary>=2.9.9,<3.0
alembic>=1.13.0,<2.0
pydantic>=2.6.0,<3.0
pydantic-settings>=2.1.0,<3.0
python-multipart>=0.0.6,<1.0
httpx>=0.27.0,<1.0
python-docx>=1.1.0,<2.0
sse-starlette>=2.0.0,<3.0
passlib[bcrypt]>=1.7.4,<2.0
python-jose[cryptography]>=3.3.0,<4.0
python-dotenv>=1.0.0,<2.0
requests>=2.31.0,<3.0
Pillow>=10.1.0,<11.0
anthropic>=0.25.0,<1.0
